--- Sources/ProcessOut/Sources/Core/Formatters/PhoneNumber/MetadataProvider/PhoneNumberMetadataProviderType.swift
@@ -1,12 +1,0 @@
-//
-//  PhoneNumberMetadataProviderType.swift
-//  ProcessOut
-//
-//  Created by Andrii Vysotskyi on 23.03.2023.
-//
-
-protocol PhoneNumberMetadataProviderType {
-
-    /// Returns metadata for given country code if any.
-    func metadata(for countryCode: String) -> PhoneNumberMetadata?
-}
